+++ internal/domain/time.go
@@ -1,15 +1,15 @@
 package domain
 
 type TimeInfo struct {
-	Start        string `json:"start_time"`
-	MachineStart string `json:"machine_start"`
-	MachineEnd   string `json:"machine_end"`
-	Quantity     int    `json:"quantity"`
-	DayQuantity  int    `json:"day_quantity"`
-	Issued       int    `json:"issued"`
-	Up           int    `json:"up" db:"up"`
-	Time         int    `json:"time" db:"time"`
-	Adjustment   int    `json:"adjustment" db:"adjustment"`
+	Start        string  `json:"start_time"`
+	MachineStart string  `json:"machine_start"`
+	MachineEnd   string  `json:"machine_end"`
+	Quantity     int     `json:"quantity"`
+	DayQuantity  int     `json:"day_quantity"`
+	Issued       int     `json:"issued"`
+	Up           int     `json:"up" db:"up"`
+	Time         float64 `json:"time" db:"time"`
+	Adjustment   int     `json:"adjustment" db:"adjustment"`
 }
 
 type TimeResult struct {
